+++ src/bosh-alicloud-cpi/mock/oss_manager_mock.go
@@ -58,3 +58,8 @@ func (a OssManagerMock) DeleteObject(bucket oss.Bucket, name string) error {
 	delete(a.mc.OssObjects, name)
 	return nil
 }
+
+func (a OssManagerMock) MultipartUploadFile(
+bucket oss.Bucket, objectKey, filePath string, partSize int64, options ...oss.Option) error {
+	return nil
+}