+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All releases of the BOSH CPI for Alibaba Cloud will be documented in this file.
 
 IMPROVEMENTS
 
-- Upgrade SDK and support HTTP PROXY ([#63](https://github.com/cloudfoundry-incubator/bosh-alicloud-cpi-release/pull/63))
+- Improve integration testcase and mock method ([#67](https://github.com/cloudfoundry-incubator/bosh-alicloud-cpi-release/pull/67))
+- Upgrade SDK and support HTTP PROXY ([#66](https://github.com/cloudfoundry-incubator/bosh-alicloud-cpi-release/pull/66))
 
 ## [r21]